+++ src/Ydb.Sdk/src/Driver.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Immutable;
-using System.Diagnostics;
 using Grpc.Core;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
@@ -9,7 +8,6 @@ using Ydb.Sdk.Ado;
 using Ydb.Sdk.Ado.Internal;
 using Ydb.Sdk.Ado.RetryPolicy;
 using Ydb.Sdk.Pool;
-using Ydb.Sdk.Tracing;
 using EndpointInfo = Ydb.Sdk.Pool.EndpointInfo;
 
 namespace Ydb.Sdk;
@@ -23,9 +21,8 @@ namespace Ydb.Sdk;
 /// </remarks>
 public sealed class Driver : BaseDriver
 {
-    private static readonly YdbRetryPolicyExecutor DiscoveryRetryPolicy = new(
-        new YdbRetryPolicy(new YdbRetryPolicyConfig { EnableRetryIdempotence = true })
-    );
+    private static readonly YdbRetryPolicyExecutor DiscoveryRetryPolicy =
+        new(YdbRetryPolicy.IdempotenceDefault, "ydb.Driver.Initialize");
 
     private readonly EndpointPool _endpointPool;
     private readonly EndpointLocalDcDetector _endpointLocalDcDetector;
@@ -71,25 +68,16 @@ public sealed class Driver : BaseDriver
     {
         Logger.LogInformation("Started initial endpoint discovery");
 
-        using var dbActivity = YdbActivitySource.StartActivity("ydb.Driver.Initialize", ActivityKind.Internal);
-        try
+        await DiscoveryRetryPolicy.ExecuteAsync(async _ =>
         {
-            await DiscoveryRetryPolicy.ExecuteAsync(async _ =>
-            {
-                await DiscoverEndpoints();
-                _discoveryTimer = new Timer(
-                    OnDiscoveryTimer,
-                    null,
-                    Config.EndpointDiscoveryInterval,
-                    Config.EndpointDiscoveryInterval
-                );
-            });
-        }
-        catch (YdbException e)
-        {
-            dbActivity?.SetException(e);
-            throw;
-        }
+            await DiscoverEndpoints();
+            _discoveryTimer = new Timer(
+                OnDiscoveryTimer,
+                null,
+                Config.EndpointDiscoveryInterval,
+                Config.EndpointDiscoveryInterval
+            );
+        });
     }
 
     private async void OnDiscoveryTimer(object? state)